# azure-devops-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Azure DevOps API",
    "description": "Connect to Azure DevOps for project and work item operations",
    "version": "1.0.0"
  },
  "host": "azuredevops-api-live.azurewebsites.net",
  "basePath": "/api",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthCheck",
        "summary": "Health check",
        "description": "Check if the API service is running",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "healthy"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                },
                "service": {
                  "type": "string",
                  "example": "Azure DevOps API Server"
                }
              }
            }
          },
          "500": {
            "description": "Service is unhealthy",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "unhealthy"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/projects": {
      "get": {
        "operationId": "ListProjects",
        "summary": "List all projects",
        "description": "Get all Azure DevOps projects",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "List of projects retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "12345678-1234-1234-1234-123456789012"
                      },
                      "name": {
                        "type": "string",
                        "example": "MyProject"
                      },
                      "description": {
                        "type": "string",
                        "example": "Project description"
                      },
                      "url": {
                        "type": "string",
                        "example": "https://dev.azure.com/organization/project"
                      },
                      "state": {
                        "type": "string",
                        "example": "wellFormed"
                      },
                      "visibility": {
                        "type": "string",
                        "example": "private"
                      }
                    }
                  }
                },
                "count": {
                  "type": "integer",
                  "example": 5
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch projects",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Authentication failed"
                }
              }
            }
          }
        }
      }
    },
    "/workitems": {
      "post": {
        "operationId": "CreateWorkItem",
        "summary": "Create work item",
        "description": "Create a new work item in Azure DevOps",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Work item data",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["project", "type", "title"],
              "properties": {
                "project": {
                  "type": "string",
                  "description": "Project name",
                  "example": "MyProject"
                },
                "type": {
                  "type": "string",
                  "description": "Work item type",
                  "enum": ["Task", "Bug", "User Story", "Feature"],
                  "example": "Task"
                },
                "title": {
                  "type": "string",
                  "description": "Work item title",
                  "example": "New task from API"
                },
                "description": {
                  "type": "string",
                  "description": "Work item description",
                  "example": "Task created via REST API"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Work item created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 123
                    },
                    "title": {
                      "type": "string",
                      "example": "New task from API"
                    },
                    "state": {
                      "type": "string",
                      "example": "New"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://dev.azure.com/organization/project/_workitems/edit/123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Project, type, and title are required"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create work item",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Authentication failed"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "GetWorkItem",
        "summary": "Get work item",
        "description": "Get work item by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Work item ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Work item details retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 123
                    },
                    "title": {
                      "type": "string",
                      "example": "Sample work item"
                    },
                    "description": {
                      "type": "string",
                      "example": "Work item description"
                    },
                    "state": {
                      "type": "string",
                      "example": "Active"
                    },
                    "assignedTo": {
                      "type": "string",
                      "example": "user@company.com"
                    },
                    "createdDate": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T10:30:00Z"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://dev.azure.com/organization/project/_workitems/edit/123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - work item ID is required",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Work item ID is required"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve work item",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Work item not found"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {},
  "parameters": {},
  "responses": {},
  "securityDefinitions": {},
  "security": [],
  "tags": []
}